# feat: Complete Phase E.2 Performance Debug File Splitting for MCP Compatibility