… .kiro/specs/performance-debug-splitting-phase-e2/api-documentation-updates.md
… .kiro/specs/performance-debug-splitting-phase-e2/design.md
… .kiro/specs/performance-debug-splitting-phase-e2/file-size-reduction-report.md
… .kiro/specs/performance-debug-splitting-phase-e2/final-validation-report.md
… .kiro/specs/performance-debug-splitting-phase-e2/main-controller-pattern-guide.md
… .kiro/specs/performance-debug-splitting-phase-e2/requirements.md
… .kiro/specs/performance-debug-splitting-phase-e2/task15-efficiency-validation-report.md
… .kiro/specs/performance-debug-splitting-phase-e2/tasks.md
… .serena/memories/development_guidelines.md
… .serena/memories/project_overview.md
… .serena/memories/suggested_commands.md
… .serena/memories/task_completion_checklist.md
… CLAUDE.md
… file-size-report.json
… src/config/FontPreloadConfig.js
… src/core/mobile-accessibility/MobileAccessibilityValidator.js
… src/debug/BenchmarkSuite.js
… src/debug/ErrorReporter.js
… src/debug/TestResultVisualizer.js
… src/debug/benchmark-suite/BenchmarkExecutor.js
… src/debug/benchmark-suite/BenchmarkReporter.js
… src/debug/benchmark-suite/BenchmarkResultAnalyzer.js
… src/debug/error-reporter/ErrorAnalyzer.js
… src/debug/error-reporter/ErrorStorage.js
… src/debug/error-reporter/ErrorSubmissionManager.js
… src/debug/test-result-visualizer/TestChartGenerator.js
… src/debug/test-result-visualizer/TestDataVisualizer.js
… src/locales/en/achievements.json
… src/locales/en/common.json
… src/locales/en/errors.json